requirements.txt
mediapipe
onnxruntime-gpu
opencv-python
rtmlib
ultralytics